Finish setup of development app
