--- example/.native/android/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = 'RNTester'
-
-include ':app'
