+++ example/native/android/keystores/debug.keystore.properties
