+++ example/native/android/settings.gradle
@@ -1,0 +1,6 @@
+rootProject.name = 'ReasonDemo'
+
+include ':app'
+
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../../../node_modules/react-native-gesture-handler/android')
