+++ package.json
@@ -6,8 +6,8 @@
     "build": "bsb -make-world",
     "start": "bsb -make-world -w",
     "clean": "bsb -clean-world -make-world",
-    "run-ios": "react-native run-ios --project-path \"./example/.native/ios\"",
-    "run-android": "react-native run-android --root \"./example/.native\"",
+    "run-ios": "react-native run-ios --project-path \"./example/native/ios\"",
+    "run-android": "react-native run-android --root \"./example/native\"",
     "start-example": "node ./node_modules/react-native/local-cli/cli.js start --root ./example"
   },
   "keywords": [
@@ -21,15 +21,15 @@
   "devDependencies": {
     "bs-platform": "^4.0.0",
     "bs-react-native": "https://github.com/reasonml-community/bs-react-native#master",
-    "reason-react": "^0.4.2",
-    "react": "^16.3.0-alpha.1",
-    "react-native": "^0.56.1",
-    "react-navigation": "^2.13.0"
+    "react": "~16.5.0",
+    "react-native": "~0.55.4",
+    "react-navigation": "^2.13.0",
+    "reason-react": "^0.5.3"
   },
   "peerDependencies": {
     "bs-react-native": "^0.9.0",
-    "react-native": "^0.55.0",
+    "react-native": "~0.55.4",
     "react-navigation": "^2.13.0",
-    "reason-react": "^0.4.2"
+    "reason-react": "^0.5.3"
   }
 }
